Реализация задания 6

+++ src/ru/neooffline/homework_j1l6/FindFood.java
@@ -1,5 +1,5 @@
 package ru.neooffline.homework_j1l6;
 
 public interface FindFood {
-    void findFood(int searchArea);
+   void findFood(double searchArea);
 }
